issue/3 - change Integer to Long for pageID's.

fix a problem when generating where pageId's are greater than Integer.MAX_VALUE by converting Integer to Long.

See merge request !1

+++ swagger-confluence-core/src/main/java/net/slkdev/swagger/confluence/config/SwaggerConfluenceConfig.java
@@ -21,7 +21,7 @@ import org.apache.commons.lang3.StringUtils;
 
 public class SwaggerConfluenceConfig {
 
-    private Integer ancestorId;
+    private Long ancestorId;
     private String authentication;
     private String confluenceRestApiUrl;
     private boolean generateNumericPrefixes;
@@ -38,11 +38,11 @@ public class SwaggerConfluenceConfig {
         paginationMode = PaginationMode.SINGLE_PAGE;
     }
 
-    public Integer getAncestorId() {
+    public Long getAncestorId() {
         return ancestorId;
     }
 
-    public void setAncestorId(final Integer ancestorId) {
+    public void setAncestorId(final Long ancestorId) {
         this.ancestorId = ancestorId;
     }
